# tmp_agent/diag/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Brain Lab",
    "version": "0.1.2"
  },
  "paths": {
...
    "/v1/agent/mission": {
      "get": {
        "summary": "Agent Get Mission",
        "description": "Devuelve misión v2 desde state/agent/<room>/mission.json.\nSi solo existe legacy, lo retorna tal cual.",
        "operationId": "agent_get_mission_v1_agent_mission_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Agent Set Mission",
        "description": "Unificado a misión v2 (mission_id/goal/status/notes) como source of truth.\nMantiene alias legacy dentro del mismo objeto para compat.",
        "operationId": "agent_set_mission_v1_agent_mission_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/agent/plan": {
      "get": {
        "summary": "Agent Get Plan",
        "operationId": "agent_get_plan_v1_agent_plan_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Agent Plan",
        "operationId": "agent_plan_v1_agent_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentPlanRequest": {
        "properties": {
          "goal": {
            "type": "string",
            "minLength": 1,
            "title": "Goal"
          },
          "room_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Id"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "goal"
        ],
        "title": "AgentPlanRequest"
      },
      "AgentPlanResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "mission": {
            "additionalProperties": true,
            "type": "object",
            "title": "Mission"
          },
          "plan": {
            "additionalProperties": true,
            "type": "object",
            "title": "Plan"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "mission",
          "plan"
        ],
        "title": "AgentPlanResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MissionCreate": {
        "properties": {
          "objective": {
            "type": "string",
            "minLength": 3,
            "title": "Objective"
          },
          "constraints": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraints"
          }
        },
        "type": "object",
        "required": [
          "objective"
        ],
        "title": "MissionCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}